Move Algolia files into the demo app.
