+++ docs/demo/bin/algolia_clear
