+++ docs/demo/bin/algolia_index
